feat: move render-error-template to own file
