+++ src/lib/functions/synchronous.js
@@ -1,10 +1,9 @@
 // @ts-check
 const { Buffer } = require('buffer')
-const { readFile } = require('fs').promises
-const { join } = require('path')
 
 const { NETLIFYDEVERR } = require('../../utils')
 
+const renderErrorTemplate = require('./render-error-remplate')
 const { detectAwsSdkError } = require('./utils')
 
 const addHeaders = (headers, response) => {
@@ -46,20 +45,6 @@ const formatLambdaLocalError = (err, acceptsHtml) =>
         trace: err.stackTrace,
       })
     : `${err.errorType}: ${err.errorMessage}\n ${err.stackTrace.join('\n ')}`
-
-let errorTemplateFile
-
-const renderErrorTemplate = async (errString) => {
-  const regexPattern = /<!--@ERROR-DETAILS-->/g
-  const templatePath = './templates/function-error.html'
-
-  try {
-    errorTemplateFile = errorTemplateFile || (await readFile(join(__dirname, templatePath), 'utf-8'))
-    return errorTemplateFile.replace(regexPattern, errString)
-  } catch {
-    return errString
-  }
-}
 
 const processRenderedResponse = async (err, request) => {
   const acceptsHtml = request.headers && request.headers.accept && request.headers.accept.includes('text/html')
